+++ src/sklearn_evaluation/plot/feature_ranking.py
@@ -22,6 +22,7 @@ from scipy.stats import shapiro
 from scipy.stats import spearmanr
 from scipy.stats import kendalltau as sp_kendalltau
 from sklearn_evaluation.telemetry import SKLearnEvaluationLogger
+from ploomber_core.exceptions import modify_exceptions
 
 import matplotlib.pyplot as plt
 
@@ -152,6 +153,7 @@ class RankD:
             self.features_ = np.array(self.features)
 
     @SKLearnEvaluationLogger.log(feature="plot")
+    @modify_exceptions
     def feature_ranks(self, X):
         """
         Parameters
@@ -171,6 +173,7 @@ class RankD:
         return self.ax
 
     @SKLearnEvaluationLogger.log(feature="plot")
+    @modify_exceptions
     def feature_ranks_custom_algorithm(self, ranks):
         """
         This method is useful if user wants to use custom algorithm for feature ranking.
@@ -187,7 +190,6 @@ class RankD:
         ax: matplotlib Axes
             Axes containing the plot
         """
-
         self._validate_rank(ranks)
         self._derive_features_from_ranks(ranks)
         self.ranks_ = ranks