+++ src/sklearn_evaluation/plot/metrics.py
@@ -2,10 +2,12 @@ import numpy as np
 from sklearn_evaluation.metrics import compute_at_thresholds
 from sklearn_evaluation.plot.util import set_default_ax
 from ..telemetry import SKLearnEvaluationLogger
+from ploomber_core.exceptions import modify_exceptions
 
 
 @set_default_ax
 @SKLearnEvaluationLogger.log(feature="plot")
+@modify_exceptions
 def metrics_at_thresholds(fn, y_true, y_score, n_thresholds=10, start=0.0, ax=None):
     """Plot metrics at increasing thresholds"""
     th, m = compute_at_thresholds(fn, y_true, y_score, n_thresholds, start)