+++ src/sklearn_evaluation/plot/validation_curve.py
@@ -1,9 +1,11 @@
 import numpy as np
 import matplotlib.pyplot as plt
 from ..telemetry import SKLearnEvaluationLogger
+from ploomber_core.exceptions import modify_exceptions
 
 
 @SKLearnEvaluationLogger.log(feature="plot")
+@modify_exceptions
 def validation_curve(
     train_scores, test_scores, param_range, param_name=None, semilogx=False, ax=None
 ):