+++ tests/test_classification_report.py
@@ -58,3 +58,15 @@ def test_raw_data_doesnt_warn(y):
     with warnings.catch_warnings():
         warnings.simplefilter("error")
         plot.ClassificationReport.from_raw_data(y_true, y_pred)
+
+
+def test_plot_classification_report_bad_input_value_error(
+    y, ploomber_value_error_message
+):
+    _, y_pred = y
+
+    with pytest.raises(ValueError, match=ploomber_value_error_message):
+        plot.ClassificationReport.from_raw_data([1, 2, 1], y_pred)
+
+    with pytest.raises(ValueError, match=ploomber_value_error_message):
+        plot.classification_report([1, 2, 1], y_pred, target_names=["Not spam", "Spam"])