+++ tests/test_confusion_matrix.py
@@ -60,3 +60,12 @@ def test_raw_data_doesnt_warn(y):
     with warnings.catch_warnings():
         warnings.simplefilter("error")
         plot.ConfusionMatrix.from_raw_data(y_true, y_pred)
+
+
+@pytest.mark.parametrize("bad_y_true_value", [None, [], [1, 2, 3, 4, 5], [1, 0], ""])
+def test_confusion_matrix_bad_input_value_error(
+    y, bad_y_true_value, ploomber_value_error_message
+):
+    _, y_pred = y
+    with pytest.raises(ValueError, match=ploomber_value_error_message):
+        plot.ConfusionMatrix(bad_y_true_value, y_pred)